Admin fully implemented v1.0

+++ new_code/admin/uploads/remove_project.php
@@ -24,6 +24,22 @@ if($project->num_rows === 1){
 
 $dir = $row[2];
 
+$fileDestination = $_SERVER['DOCUMENT_ROOT'] . '/projects/vava_project/imgs/' . $dir . '/';
+$it = new RecursiveDirectoryIterator($fileDestination, RecursiveDirectoryIterator::SKIP_DOTS);
+$files = new RecursiveIteratorIterator($it, RecursiveIteratorIterator::CHILD_FIRST);
+
+foreach($files as $file) {
+    if ($file->isDir()){
+        rmdir($file->getRealPath());
+    } else {
+        unlink($file->getRealPath());
+    }
+}
+
+if (!rmdir($fileDestination)) {
+	die("Directory failed to be removed.");
+}
+
 // Get the new values
 
 $sql  = "DELETE FROM `photography_projects` WHERE `photography_projects`.`project_id` = " . $_POST["id"];
@@ -32,7 +48,7 @@ if(!mysqli_query($con, $sql)){
   die("ERROR: Not able to execute $sql." . mysqli_error($sql));
 }
 
-header('Location: ../editproject.php?project=' . $_POST['id']);
+header('Location: ../existingprojects.php');
 
 $con->close();
 // If the user is not logged in redirect to the login page...
